requirements.txt
flask
flask-cors
google-generativeai
youtube-transcript-api
beautifulsoup4
redis
requests
gunicorn